test/example_thirdparty/requirements.txt
South>=0.7.5
Markdown>=2.1
postmarkup>=1.2.0

# TODO: remove dependencies
django-annoying>=0.7.6
django-pure-pagination>=0.2.1
sorl-thumbnail>=11.12
pytils>=0.2.3

# Theme, templates and authentication
django-simple-captcha
pinax-theme-bootstrap==2.0.4
django-forms-bootstrap>=2.0.3
git+https://github.com/pinax/django-user-accounts.git
git+https://github.com/pinax/pinax-theme-bootstrap-account.git
